requirements.txt
pillow
torch
opencv-python
numpy
pandas
faiss-cpu
transformers
streamlit
flask
flask-cors